CSSO & UglifyJS injection (without tests)
