+++ lib/borschik.js
@@ -36,18 +36,21 @@ module.exports = require('coa').Cmd()
             return (f == 'yes' || f == 'no')? f : 'yes';
         })
         .end()
+    .opt()
+        .name('minimize').title('Minimize')
+        .short('m').long('minimize')
+        .def('yes')
+        .val(function(m) {
+            return (m == 'yes' || m == 'no')? m : 'yes';
+        })
+        .end()
     .act(function(opts, args, res) {
-        var t = opts.tech,
-            tech;
+        var t = opts.tech;
 
         t = PATH.basename(t) === t ?
                 PATH.join(__dirname, 'techs', t + '.js') :
                 PATH.resolve(t);
 
-        tech = new (require(t).Tech)(opts);
-
-        return tech
-                .createFile(opts.input)
-                .read()
-                .write(opts.output);
+        return new (require(t).Tech)(opts)
+            .process(opts.input, opts.output);
     });
