+++ lib/tech.js
@@ -17,6 +17,15 @@ exports.Tech = INHERIT({
             return file;
         },
 
+        process: function(path, out) {
+            var res = this.createFile(path, 'include').process();
+            return U.writeFile(out, this.opts.minimize == 'yes' ? this.minimize(res) : res);
+        },
+
+        minimize: function(content) {
+            return content;
+        },
+
         File: exports.File = INHERIT({
 
             __constructor: function(tech, path, type, parent) {
@@ -50,40 +59,7 @@ exports.Tech = INHERIT({
             },
 
             write: function(output) {
-
-                var res = this.process(this.path);
-
-                // save res to file
-                if (typeof output === 'string') {
-                    return QFS.write(output, res);
-                }
-
-                // output res to stdout
-                if (output === process.stdout) {
-                    output.write(res);
-                    return Q.resolve();
-                }
-
-                // write res to writable stream of opened file
-                var defer = Q.defer();
-
-                output.on('error', function(err) {
-                    defer.reject(err);
-                });
-
-                output.once('close', function() {
-                    defer.resolve();
-                });
-
-                output.once('end', function() {
-                    defer.resolve();
-                });
-
-                output.write(res);
-                output.end();
-
-                return defer.promise;
-
+                return U.writeFile(output, this.process(this.path));
             },
 
             process: function(path) {
