+++ lib/techs/css-base.js
@@ -2,9 +2,15 @@ var INHERIT = require('inherit'),
     FREEZE = require('../freeze'),
     util = require('util'),
     base = require('../tech'),
-    PATH = require('path');
+    PATH = require('path'),
+    CSSO = require('csso');
 
 exports.Tech = INHERIT(base.Tech, {
+
+    minimize: function(content) {
+        return CSSO.justDoIt(content);
+    },
+
     File: exports.File = INHERIT(base.File, {
 
         processLink: function(path) {
