+++ lib/techs/css.js
@@ -42,12 +42,12 @@ exports.Tech = INHERIT(cssbase.Tech, {
         },
 
         processInclude: function(path) {
-          function SpecificTranslator(source) {
-              translator.call(this, source);
-              this._path = path;
-          };
-          return correctInherit(SpecificTranslator, translator)
-              .translate(path, this.content);
+            function SpecificTranslator(source) {
+                  translator.call(this, source);
+                  this._path = path;
+            };
+
+            return correctInherit(SpecificTranslator, translator).translate(path, this.content);
         },
 
         processPath: function(path) {
