+++ lib/techs/js.js
@@ -1,8 +1,13 @@
 var INHERIT = require('inherit'),
     base = require('../tech'),
-    FS = require('../fs');
+    FS = require('../fs'),
+    UJS = require('uglify-js');
 
 exports.Tech = INHERIT(base.Tech, {
+
+    minimize: function(content) {
+        return require('uglify-js')(content);
+    },
 
     File: exports.File = INHERIT(base.File, {
