+++ package.json
@@ -23,7 +23,9 @@
     "q": "~0.8.10",
     "q-fs": "0.1",
     "cssp": "~1.0.6",
-    "coffee-script": "*"
+    "coffee-script": "*",
+    "csso": "~1.3.6",
+    "uglify-js": "~1.3.0"
   },
   "devDependencies": {
     "istanbul": "~0.1.22",
